refactor: extract browser-action badge logic and readability fonts
Move updateIcon/getIsReadabilityActive out of styles.ts into a
dedicated background/badge.ts, and move readabilityFonts out of the
generic settings module into readability/fonts.ts, since both were
readability-specific concerns living in unrelated modules.
