+++ src/editor/components/magic/readability/TheReadabilityFontFamily.vue
@@ -16,8 +16,7 @@
 import Vue from 'vue';
 
 import { StylebotFonts } from '@stylebot/types';
-import { readabilityFonts } from '@stylebot/settings';
-import { previewReaderFont } from '@stylebot/readability';
+import { readabilityFonts, previewReaderFont } from '@stylebot/readability';
 
 import FontFamilyDropdown from '../../text/FontFamilyDropdown.vue';
