+++ src/readability/fonts.ts
@@ -1,0 +1,10 @@
+import { StylebotFonts } from '@stylebot/types';
+
+// Kept apart from `defaultOptions.fonts`, which also powers the
+// general-purpose Text panel and needs display/code fonts too.
+export const readabilityFonts: StylebotFonts = [
+  'Merriweather',
+  'Georgia',
+  'Lora',
+  'Helvetica',
+];
