+++ src/readability/index.ts
@@ -3,3 +3,4 @@ import './index.scss';
 export { apply, remove } from './apply';
 export { isReaderable } from './heuristics';
 export { previewReaderFont, onPreviewReaderFont } from './preview';
+export { readabilityFonts } from './fonts';
